Dashboard and submenu fixes
Saved submenu now with date section breaks, Dashboard redesign, and
fixed most recent download datetime string

+++ src/org/odk/clinic/android/activities/ViewCompletedForms.java
@@ -177,7 +177,7 @@ public class ViewCompletedForms extends ListActivity {
 						//Submitted on vs. Saved on ... etc.
 						form.setDisplaySubtext(c.getString(displaySubtextIndex));
 						//used for splitting into date menus
-						form.setDate((int) c.getLong(dateIndex));
+						form.setDate(c.getLong(dateIndex));
 						mTotalForms.add(form);
 					}
 				} while (c.moveToNext());
